+++ R/mod_tabpanel_input.R
@@ -19,15 +19,20 @@ mod_tabpanel_input_ui <- function(id) {
       height = "100px",
       width = "100px"
     ),
-    shiny::uiOutput(ns("input_tab_ui"))
+    shiny::div(
+      class = "content-container",
+      shiny::div(
+        class = "card-container",
+        shiny::uiOutput(ns("input_tab_ui"))
+      ),
+      footer_text
+    )
   )
 }
 
 
 #' tabpanel "input" Server Functions
-#' @param id,input,output,session standard \code{shiny} boilerplate
-#' @param data reactive input dataset preprocessed if no errors
-#' @param errors_detected reactive boolean, when TRUE errors on mandatory variables where detected
+#' @noRd
 mod_tabpanel_input_server <- function(id, data, errors_detected) {
   shiny::moduleServer(id, function(input, output, session) {
     ns <- session$ns
@@ -43,88 +48,64 @@ mod_tabpanel_input_server <- function(id, data, errors_detected) {
     # ensuring that content is onlyu shown if data check returns no errors
     output$input_tab_ui <- shiny::renderUI({
       if (errors_detected() == TRUE) {
-        return(datacheck_error_message)
+        datacheck_error_message
       } else {
-        return(shiny::tagList(
-          tags$style(
-            HTML("
-      .well {
-        margin-top: 20px; /* Adjust the value to move the wellPanel lower */
-      }
-    ")
-          ),
-          shiny::fluidRow(
-            column(
-              6,
-              shiny::wellPanel(
-                span("Dataset settings", style = "font-size:140%;color:#304794"),
-                hr(),
-                span("Pathogen", style = "font-size:100%;font-weight: bold"),
-                shiny::uiOutput(ns("pathogen_choices")),
-                br(),
-                span("Filters", style = "font-size:100%;font-weight: bold"),
-                br(),
-                span("You can chose to investigate a subset of your data according to the filters you select. When filtering by date_report you have the possibility select a specific timeperiod you want to investigate. In the timeseries visualisation only the timeperiod you selected will be shown and the outbreak detection algorithms will only train on the data from the timeperiod you selected."),
-                br(),
-                br(),
-                shiny::div(
-                  id = "filter_input",
-                  span(
-                    "Add and remove filters",
-                    shiny::actionButton(
-                      inputId = ns("add_filter"),
-                      label = "",
-                      icon = shiny::icon("plus")
-                    ),
-                    shiny::actionButton(
-                      inputId = ns("remove_filter"),
-                      label = "",
-                      icon = shiny::icon("minus")
-                    )
-                  ),
-                ),
-                br(),
-                mod_input_filter_ui(id = ns("filter0"))
+        bslib::layout_columns(
+          col_widths = c(6, 6),
+          bslib::card(
+            bslib::card_title("Dataset settings", container = shiny::h1),
+            shiny::h2("Pathogen"),
+            shiny::uiOutput(ns("pathogen_choices")),
+            shiny::br(),
+            shiny::h2("Filters"),
+            span("You can chose to investigate a subset of your data according to the filters you select. When filtering by date_report you have the possibility select a specific timeperiod you want to investigate. In the timeseries visualisation only the timeperiod you selected will be shown and the outbreak detection algorithms will only train on the data from the timeperiod you selected."),
+            shiny::div(
+              id = "filter_input",
+              shiny::div(
+                "Add and remove filters",
+              ),
+              shiny::actionButton(
+                inputId = ns("add_filter"),
+                label = "",
+                icon = shiny::icon("plus")
+              ),
+              shiny::actionButton(
+                inputId = ns("remove_filter"),
+                label = "",
+                icon = shiny::icon("minus")
               )
             ),
-            column(
-              6,
-              shiny::wellPanel(
-                span("Signal Detection settings", style = "font-size:140%;color:#304794"),
-                hr(),
-                span("Strata", style = "font-size:100%;font-weight: bold"),
-                br(),
-                span("Select up to 3 variables you want to stratify by. Signals and visualisations will be generated for each stratum."),
-                shiny::uiOutput(ns("strat_choices")),
-                br(),
-                span("Signal Detection Period", style = "font-size:100%;font-weight: bold"),
-                br(),
-                span("Set the number of weeks you want to generate signals for. The signals are generated for the most recent weeks."),
-                br(),
-                shiny::uiOutput(ns("weeks_selection")),
-                shiny::textOutput(ns("text_weeks_selection")),
-                br(),
-                shiny::fluidRow(
-                  column(
-                    6,
-                    span("Signal detection algorithm", style = "font-size:100%;font-weight: bold"),
-                    br(),
-                    span("Depending on the number of weeks you want to generate signals for and the filters you set, the choice of algorithms is automatically updated to those which are possible to apply for your settings."),
-                    br(),
-                    shiny::uiOutput(ns("algorithm_choice"))
-                  ),
-                  column(
-                    6, shiny::conditionalPanel(
-                      condition = sprintf("output['%s'] == 'TRUE'", ns("algorithm_glm")),
-                      checkboxInput(ns("pandemic_correction"), "Covid19 Pandemic Correction", value = FALSE)
-                    ),
-                    shiny::uiOutput(ns("conditional_date_input"))
-                  )
-                )
+            mod_input_filter_ui(id = ns("filter0"))
+          ),
+          bslib::card(
+            bslib::card_title("Signal Detection settings", container = shiny::h1),
+            shiny::h2("Strata"),
+            shiny::span("Select up to 3 variables you want to stratify by. Signals and visualisations will be generated for each stratum."),
+            shiny::uiOutput(ns("strat_choices")),
+            shiny::br(),
+            shiny::h2("Signal Detection Period"),
+            shiny::span("Set the number of weeks you want to generate signals for. The signals are generated for the most recent weeks."),
+            shiny::uiOutput(ns("weeks_selection")),
+            shiny::textOutput(ns("text_weeks_selection")),
+            shiny::br(),
+            shiny::fluidRow(
+              shiny::column(
+                width = 6,
+                shiny::h2("Signal detection algorithm"),
+                shiny::span("Depending on the number of weeks you want to generate signals for and the filters you set, the choice of algorithms is automatically updated to those which are possible to apply for your settings."),
+                shiny::uiOutput(ns("algorithm_choice"))
+              ),
+              shiny::column(
+                width = 6,
+                shiny::conditionalPanel(
+                  condition = sprintf("output['%s'] == 'TRUE'", ns("algorithm_glm")),
+                  checkboxInput(ns("pandemic_correction"), "Covid19 Pandemic Correction", value = FALSE)
+                ),
+                shiny::uiOutput(ns("conditional_date_input"))
               )
             )
           )
-        ))
+        )
       }
     })
 
@@ -180,7 +161,7 @@ mod_tabpanel_input_server <- function(id, data, errors_detected) {
       shiny::req(!errors_detected())
       return(shiny::selectInput(
         inputId = ns("pathogen_vars"),
-        label = "",
+        label = "Select a pathogen",
         choices = unique(data()$pathogen),
         width = "40%"
       ))
@@ -337,7 +318,7 @@ mod_tabpanel_input_server <- function(id, data, errors_detected) {
 
 
     output$strat_choices <- shiny::renderUI({
-      req(!errors_detected())
+      shiny::req(!errors_detected())
       shiny::req(available_var_opts)
 
       shiny::selectizeInput(
@@ -529,13 +510,10 @@ mod_tabpanel_input_server <- function(id, data, errors_detected) {
       }
     })
 
-
     # get default, min and max dates for intervention date
     valid_dates_intervention <- shiny::reactive({
       get_valid_dates_intervention_start(filtered_data(), number_of_weeks = input$n_weeks, time_trend = time_trend())
     })
-
-    # output$valid_dates_intervention <-
 
     # Return list of subsetted data and parameters
     return(list(